backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-migrate
flask-cors
flask-jwt-extended
python-dotenv
werkzeug
psycopg2-binary
gunicorn
requests
pytest
pytest-flask
bandit

